Removed the unsupported `redirect_uri` parameter.
Stripe does not consider the `redirect_uri` parameter as valid when it's sent to the /oauth/authorize connect endpoint. Currently, they're ignoring this parameter, but have plans to support it in the future. However, as of September 30, 2013, they will start returning an error on invalid `redirect_uri` parameters sent to their connect endpoint.